make better structure
