+++ src/components/Login/Login.css
@@ -1,7 +1,8 @@
-.btn-group-vertical {
-    left: 50%;
+#login {
+	position: fixed;
+	top: 45%;
+	left: 50%;
     transform: translate(-50%, -50%);
-    top: 400px;
 }
 
 .textfield {
